rename style files
